- Created a censor_password processor to censor eventual passwords passed to structlog - abstracted Exception Response to a function return_exception_response - Renamed log folder to logging
