+++ api/core/exceptions/handlers.py
@@ -1,12 +1,27 @@
 # -*- coding: utf-8 -*-
+from typing import Any, Dict, Optional
+
 from fastapi import Request, status
 from fastapi.responses import JSONResponse
 
-from api.core.log.caishen_logger import logger
+from api.core.logging.caishen_logger import logger
 
 from .authentication import AuthenticationProviderException
 from .authorization import UnAuthorizedUser
 from .database import DatabaseConnectionError
+
+
+async def return_exception_response(
+    detail: str, status_code: int, headers: Optional[Dict[str, Any]] = None
+) -> JSONResponse:
+    if headers:
+        return JSONResponse(
+            {"detail": detail},
+            status_code=status_code,
+            headers=headers,
+        )
+    else:
+        return JSONResponse({"detail": detail}, status_code=status_code)
 
 
 async def database_connection_error_handler(
@@ -15,17 +30,11 @@ async def database_connection_error_handler(
 ) -> JSONResponse:
     logger.exception("Failed to connect to the database", repr(exc))
     headers = getattr(exc, "headers", None)
-    if headers:
-        return JSONResponse(
-            {"detail": "An error has occurred. Please try again."},
-            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-            headers=headers,
-        )
-    else:
-        return JSONResponse(
-            {"detail": "An error has occurred. Please try again."},
-            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-        )
+    return await return_exception_response(
+        detail="An error has occurred. Please try again.",
+        status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
+        headers=headers,
+    )
 
 
 async def unauthorized_user_handler(
@@ -33,17 +42,11 @@ async def unauthorized_user_handler(
 ) -> JSONResponse:
     logger.warning(**exc.log.dict())
     headers = getattr(exc, "headers", None)
-    if headers:
-        return JSONResponse(
-            {"detail": exc.log.msg},
-            status_code=exc.log.status_code,
-            headers=headers,
-        )
-    else:
-        return JSONResponse(
-            {"detail": exc.log.msg},
-            status_code=exc.log.status_code,
-        )
+    return await return_exception_response(
+        detail=exc.log.msg,
+        status_code=exc.log.status_code,
+        headers=headers,
+    )
 
 
 async def authentication_provider_error_handler(
@@ -52,14 +55,8 @@ async def authentication_provider_error_handler(
 ) -> JSONResponse:
     logger.error(**exc.log.dict())
     headers = getattr(exc, "headers", None)
-    if headers:
-        return JSONResponse(
-            {"detail": exc.log.msg},
-            status_code=exc.log.status_code,
-            headers=headers,
-        )
-    else:
-        return JSONResponse(
-            {"detail": exc.log.msg},
-            status_code=exc.log.status_code,
-        )
+    return await return_exception_response(
+        detail=exc.log.msg,
+        status_code=exc.log.status_code,
+        headers=headers,
+    )
